[move/#243] 이미지 압축 로직 registration -> presigned-url 모듈로 이동
