+++ data/registration/src/main/java/com/napzak/market/registration/repository/RegistrationRepositoryImpl.kt
@@ -1,17 +1,14 @@
 package com.napzak.market.registration.repository
 
-import androidx.core.net.toUri
 import com.napzak.market.registration.datasource.RegistrationDataSource
 import com.napzak.market.registration.mapper.toData
 import com.napzak.market.registration.mapper.toDomain
 import com.napzak.market.registration.model.PurchaseRegistrationProduct
 import com.napzak.market.registration.model.SaleRegistrationProduct
-import com.napzak.market.remote.ImageCompressor
 import javax.inject.Inject
 
 class RegistrationRepositoryImpl @Inject constructor(
     private val registrationDataSource: RegistrationDataSource,
-    private val imageCompressor: ImageCompressor,
 ) : RegistrationRepository {
     override suspend fun postSaleRegistration(
         saleRegistrationProduct: SaleRegistrationProduct,
@@ -56,17 +53,5 @@ class RegistrationRepositoryImpl @Inject constructor(
     ): Result<Unit> = runCatching {
         registrationDataSource
             .putPurchaseRegistration(productId, purchaseRegistrationProduct.toData())
-    }
-
-    override suspend fun compressProductImage(
-        imageUri: String,
-    ): Result<String> = runCatching {
-        imageCompressor.compressImage(imageUri.toUri())
-    }.mapCatching { uri ->
-        uri.toString()
-    }
-
-    override fun clearCachedImage(): Result<Unit> = runCatching {
-        imageCompressor.clearCachedImage()
     }
 }
