+++ domain/registration/src/main/java/com/napzak/market/registration/repository/RegistrationRepository.kt
@@ -29,10 +29,4 @@ interface RegistrationRepository {
         productId: Long,
         purchaseRegistrationProduct: PurchaseRegistrationProduct,
     ): Result<Unit>
-
-    suspend fun compressProductImage(
-        image: String,
-    ): Result<String>
-
-    fun clearCachedImage(): Result<Unit>
 }
