--- domain/registration/src/main/java/com/napzak/market/registration/usecase/ClearCacheUseCase.kt
@@ -1,11 +1,0 @@
-package com.napzak.market.registration.usecase
-
-import com.napzak.market.registration.repository.RegistrationRepository
-import javax.inject.Inject
-
-class ClearCacheUseCase @Inject constructor(
-    private val registrationRepository: RegistrationRepository,
-) {
-    operator fun invoke(): Result<Unit> =
-        registrationRepository.clearCachedImage()
-}
