+++ feature/registration/src/main/java/com/napzak/market/registration/purchase/PurchaseRegistrationViewModel.kt
@@ -16,8 +16,8 @@ import com.napzak.market.registration.model.Photo
 import com.napzak.market.registration.model.ProductImage
 import com.napzak.market.registration.model.PurchaseRegistrationProduct
 import com.napzak.market.registration.purchase.state.PurchaseContract.PurchaseUiState
-import com.napzak.market.registration.usecase.ClearCacheUseCase
-import com.napzak.market.registration.usecase.CompressImageUseCase
+import com.napzak.market.presigned_url.usecase.ClearCacheUseCase
+import com.napzak.market.presigned_url.usecase.CompressImageUseCase
 import com.napzak.market.registration.usecase.EditRegisteredProductUseCase
 import com.napzak.market.registration.usecase.GetRegisteredPurchaseProductUseCase
 import com.napzak.market.registration.usecase.RegisterProductUseCase
